requirements.txt
Flask==2.0.1
tensorflow>=2.5.0*
keras==2.4.3
numpy==1.19.5
opencv-python==4.5.1.*
matplotlib==3.4.1
pillow==8.2.0
pafy==0.5.5
youtube_dl==2021.6.6
